notes view
