+++ Noted.Api/Noted.Api/Data/Note.cs
@@ -1,12 +1,20 @@
-﻿namespace Noted.Api.Data
+﻿using System.ComponentModel.DataAnnotations.Schema;
+
+namespace Noted.Api.Data
 {
     public class Note
     {
         public Guid Id { get; set; }
         public string Title { get; set; } = "";
         public string Content { get; set; } = "";
+
         public DateTime Created { get; set; }
         public bool deleted { get; set; } 
+
+        public Note()
+        {
+            this.Created = DateTime.UtcNow;
+        }
 
     }
 }
