+++ client/src/router/index.ts
@@ -1,5 +1,6 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import HomeViewVue from '@/views/HomeView.vue'
+import NoteEditorVue from '@/components/NoteEditor.vue'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -8,6 +9,11 @@ const router = createRouter({
       path: '/',
       name: 'home',
       component: HomeViewVue
+    },
+    {
+      path: '/note-editor',
+      name: 'NodeEditor',
+      component: NoteEditorVue
     }
   ]
 })
