+++ client/src/views/NoteEditView.vue
@@ -1,0 +1,6 @@
+<template>
+  <NoteEditor></NoteEditor>
+</template>
+<script setup lang="ts">
+import NoteEditor from '../components/NoteEditor.vue'
+</script>
